--- a/src/hts/hts_engine_API/lib/HTS_hidden.h
+++ b/src/hts/hts_engine_API/lib/HTS_hidden.h
@@ -494,6 +494,7 @@ typedef struct _HTS_Vocoder {
    size_t lsp2lpc_size;         /* buffer size of lsp2lpc */
    double *gc2gc_buff;          /* used in gc2gc */
    size_t gc2gc_size;           /* buffer size for gc2gc */
+   int d2offset;
 } HTS_Vocoder;
 
 /* HTS_Vocoder_initialize: initialize vocoder */