… src/hts/hts_engine_API/lib/HTS_vocoder.c