Replace the mlsa filter originally from flite with the hand optimized
ones in bell bird.

# Preliminary benchmark

Using the `docs/alice` text.

## With Bellbird optimizations

real    1m44,649s
user    1m43,712s
sys     0m0,696s

## Without:

real    3m47,627s
user    3m46,860s
sys     0m0,732s