@zeehio 2017-08-15
I am sure you checked and Peter did as well in Bellbird but just to be sure... Are the generated audio files equal?

Quite impressive to see how good human optimized code is with respect to letting the compiler do the optimizations!