@forslund 2017-08-15
I didn't expect them to be, but according to diff they're identical. Please try it out if you have time after you get back from the _playa_