@zeehio 2017-08-15
Could you please add the support as well to hts voices? 
https://github.com/peterdrysdale/bellbird/blob/master/src/hts/HTS_vocoder.c

Looks mergeable to me :-)