@penrods 2017-08-15
Wow, those are some impressive optimizations!