@penrods 2017-08-28
Just curious -- is this on hold?  I glanced at the CI log and it appears the win-shared build just stalled, not a definitive failure.