@zeehio 2017-08-28
The PR is fine, I believe @forslund was going to add the optimization as well to the HTS voices, so both voice techniques benefit from this optimization. This looks good to be merged to me, I leave it up to him to decide if he wants to merge now or wait for later 👍 . In any case, we are splitting the mimic repo in several smaller repositories, both for practical reasons (smaller repository size, faster compilation times) and for licensing reasons (the Spanish language support is GPL due to the Saga library dependency). Due to this split, this change will have to go to the https://github.com/MycroftAI/mimic-core repository and the HTS patch will need to go to the https://github.com/MycroftAI/htsengine_meson