@forslund 2017-08-29
The hts was more different than the clustergen than I originally thought so I didn't quite finish and then I got sidetracked. Will fix it today/tomorrow and then move the changes over to the new system.