@forslund 2017-08-31
Finally! Been battling for a day with a memory issue and have finally found it. Cleanup and then maybe HTS will be ready... unless I mess things up even more.