@zeehio 2017-08-31
I'm checking the PR from my phone and it looks great and clean! Just a couple of comments:

The Bellbird optimization only works for a pade order of 5. For that reason the `SPEED_HACK` #define in clustergen voices is correctly removed in this pull request. HTS voices have the equivalent `HTS_EMBEDDED` #define at https://github.com/forslund/mimic/blob/4a9ad5ae169e3936c2f3d165c155ac77187f7d60/src/hts/hts_engine_API/lib/HTS_hidden.h#L447 . Maybe this `HTS_EMBEDDED` define should be removed as well, as the Bellbird optimization does not work if the pade order is not 5. I don't recall the effect of `GAUSS` and `IRLENG` so I don't know if they are related to the Pade order in any way. Feel free to either check if they are compatible with the Bellbird optimization or more simply drop the whole HTS_EMBEDDED block.

The second comment is not code-related: could you please add the new file to the COPYING file? And maybe add Peter Drysdale and the Bellbird project to the ACKNOWLEDGEMENTS file as well?

Thanks for all this work!