@forslund 2017-08-31
Of course, will fix this.

Should have thought of adding him to acknowledgements, the result of the optimizations are amazing