@forslund 2017-09-01
As far as I can determine both GAUSS and IRLEN does some optimization and still works so I've kept those.

I didn't find any copyright period of bellbird, not sure if that's an issue in the COPYING file.