@zeehio 2017-09-06
The year in the copyright is used to determine when the copyright will expire. A work that has not done any copyrightable change in 200 years (well, much less than 200, but it depends on the legislation) is public domain.
 
The safest thing for you to do when you receive a work without copyright year is to assume the last change was done the day you received the work. Therefore here you could use 2017 safely.

Let's merge this now. It is good!